Mutate Game's entities to "our domain" entities
The game is not written in typescript so all our classes and heritage
and what not meens _nothing_ to The Game.

Checks are made on object sent to Game's functions and methods that
doesn't identify our _subclasses_ as valid instances for said methods.

Game's documentation shows us that _Modifying object prototypes_ is a
valid technique to address our needs ( see
https://docs.screeps.com/contributed/modifying-prototypes.html ).

We have decided to group those prototype changes inside something called
a _MutationFacility_.

Co-authored-by: Patrick P. Henley <patrick.p.henley@gmail.com>